Revert "refactor(cleanup): Apply coding rules (#1499)"
This reverts commit 4cc68bb5245fa08e22f0a12c8d4b73192b86aef2.

+++ engine/src/main/java/org/operaton/bpm/engine/impl/persistence/entity/HistoricVariableInstanceManager.java
@@ -94,8 +94,8 @@ public class HistoricVariableInstanceManager extends AbstractHistoricManager {
       List <HistoricVariableInstanceEntity> cachedHistoricVariableInstances = getDbEntityManager().getCachedEntitiesByType(HistoricVariableInstanceEntity.class);
       for (HistoricVariableInstanceEntity historicVariableInstance : cachedHistoricVariableInstances) {
         // make sure we only delete the right ones (as we cannot make a proper query in the cache)
-        if ((historicProcessInstanceId.equals(historicVariableInstance.getProcessInstanceId()))
-            || (historicCaseInstanceId.equals(historicVariableInstance.getCaseInstanceId()))) {
+        if ((historicProcessInstanceId != null && historicProcessInstanceId.equals(historicVariableInstance.getProcessInstanceId()))
+            || (historicCaseInstanceId != null && historicCaseInstanceId.equals(historicVariableInstance.getCaseInstanceId()))) {
           historicVariableInstance.delete();
         }
       }
